move wave_geom stuff to RectCart.lua
